alpinetest/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.8
django-admin-honeypot==1.1.0
django-ckeditor==6.1.0
django-embed-video==1.4.0
django-js-asset==1.2.2
idna==3.2
Pillow==8.3.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
psycopg2-binary==2.9.1
gunicorn==20.1.0
